refactor: 플랫폼 sync mutation을 usePlatformSyncMutations 훅으로 분리
